+++ src/demos/Demos.Collisions.Interactable/Services/Ui/AlgorithmSelectWindow.cs
@@ -1,4 +1,5 @@
-﻿using Detach;
+﻿using Demos.Collisions.Interactable.Services.States;
+using Detach;
 using Detach.Collisions;
 using Detach.Collisions.Primitives2D;
 using Detach.Collisions.Primitives3D;
