+++ src/internals/CollisionFormats/Serialization/CollisionScenario.cs
@@ -4,9 +4,9 @@ public sealed record CollisionScenario<TParam1, TParam2>
 	where TParam1 : struct
 	where TParam2 : struct
 {
-	public CollisionScenario(string algorithmName, TParam1[] params1, TParam2[] params2)
+	public CollisionScenario(string algorithmName, List<TParam1> params1, List<TParam2> params2)
 	{
-		if (params1.Length != params2.Length)
+		if (params1.Count != params2.Count)
 			throw new ArgumentException("Parameter arrays must be the same length.");
 
 		AlgorithmName = algorithmName;
@@ -15,6 +15,6 @@ public sealed record CollisionScenario<TParam1, TParam2>
 	}
 
 	public string AlgorithmName { get; init; }
-	public TParam1[] Params1 { get; init; }
-	public TParam2[] Params2 { get; init; }
+	public List<TParam1> Params1 { get; init; }
+	public List<TParam2> Params2 { get; init; }
 }
